Create common constructor for Opaque types
The constructors takes the handle (IntPtr) and ownership (Transfer) as arguments. In the case of ReferenceCountedOpaques, we keep a table of instances so that we return the same instance instead of a new instance if we have already marshaled it once.

Other Opaques are treated more like value types and will get a new instance very time.

Dropped the special Copy method from OwnedOpaques so that it does not have to be a generic type.

Moved Transfer Enum from CodeGen.Model to Core.

Got rid of CustomMarshaler stuff.

In generated code, constructor implementations are now done in a static method that returns the handle. This way, we can just pass the handle to the common constructor.

Added NullHandleIsInstanceAttribute to flag classes where Handle == IntPtr.Zero is an instance rather than a null object.