Fix error when creating new instance
This fixes an exception when trying to create a new instance of an object
that is still in the object map but is no longer alive because it is not
referenced but has not been finalized.

This was causing a test to fail intermitantly because it was trying to create a
new instance of a ParamSpec.