Fix potential concurrency issue
This fixes a potiental issue where an we could try to get the same object at the same
time and end up throwing an exception in the constructor because it is already in the
object map.

The problem could happen if both threads check for an existing instance before creating
the new instance. By wrapping the check and the create in a lock block, we prevent this
from happening.