get rid of string overloads
These had the problem that parameters types could not be mixed. It was
either all string or all utf8.

It is replaced with implicit casts. This requires a ConditionalWeakTable
for proper memory management, otherwise the temporary Utf8 object could
be garbage collected before the unowned ref struct is used. This has a
nice side effect of caching string conversions.