use .gir for depedency tree
since we are now saving the fixed up .gir files, we can use those for depedency resolution instead of trying to reverse engineer the type info from .NET reflection.

Callbacks were a particular problem since there is a significant amount of metadata needed to give all of the information for the unmanaged version of the callback.